Renames composable files & simplifies useMq syntax
